Add BFSearch method to Graph class and knightMoves wrapper
The .BFSearch([row1,col1],[row2,col2]) method has been added
to the Graph class to find the shortest path between two
nodes of a graph.

Breadth-first search has been used, since  the graph is
unweighted by construction. Indeed, all the possible paths
starting from the source node are searched, with growing
lengths, until the destination node is reached.

If the destination node is not found, or the source /
destinationan nodes do not exist, an empty array is returned.

Note: as the graph is unweighted, currently nodes already
visited can potentially be visited again, creating cycles.
The shortest path cannot be a cycle (because, excluding this
cycle, a shorter path can be found), hence such paths are
never returned, if one exists. However, this might increase
the computautional time. Then, the next step to be done is
excluding the already visited nodes in the search process.

A wrapper method .knightMoves([row1,col1],[row2,col2]) has
been added to the KnightTravails class. Some formatted output
is returned based on the outcome of the search.

This .knightMoves() is the actual function which is required
by this project. Instead of having a simple function, using
an object allows us to precompute only one the moves graph and
store it in the private properties of a KnightTravails object.